Move transfer component setup to be with other transfer component code
