+++ plugin-hrm-form/src/HrmFormPlugin.tsx
@@ -43,10 +43,11 @@ import { setUpReferrableResources } from './components/resources/setUpReferrable
 import { subscribeNewMessageAlertOnPluginInit } from './notifications/newMessage';
 import { subscribeReservedTaskAlert } from './notifications/reservedTask';
 import { setUpCounselorToolkits } from './components/toolkits/setUpCounselorToolkits';
-import { setupConferenceComponents, setUpConferenceActions } from './conference';
+import { setUpConferenceActions, setupConferenceComponents } from './conference';
 import { setUpTransferActions } from './transfer/setUpTransferActions';
 import { playNotification } from './notifications/playNotification';
 import { namespace } from './states/storeNamespaces';
+import { setUpTransferComponents } from './components/transfer/setUpTransferComponents';
 
 const PLUGIN_NAME = 'HrmFormPlugin';
 
@@ -110,7 +111,7 @@ const setUpComponents = (
   }
 
   if (featureFlags.enable_transfers) {
-    Components.setUpTransferComponents();
+    setUpTransferComponents();
     Channels.setUpIncomingTransferMessage();
   }
